+++ CHANGELOG.md
@@ -1565,3 +1565,4 @@
 - [2026-05-28 22:25] feat(forms): integrate custom telemetry logger in forms
 - [2026-05-28 22:27] feat(navigation): integrate custom telemetry logger in navigation
 - [2026-05-28 23:21] style(analytics): adjust design tokens, typography, and contrast in analytics
+- [2026-05-29 09:03] test(navigation): add unit tests for navigation business logic